packages/reflex-chess-viewer/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reflex-chess-viewer"
version = "0.1.0"
description = "Reference viewer: chessboard + notation + PGN to PackedGameTree"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.10"
authors = [{ name = "kuruhuru" }]
dependencies = [
  "reflex>=0.8.23",
  "python-chess>=1.999",
  "reflex-chess-model",
  "reflex-chessboard",
  "reflex-chess-notation",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
reflex_chess_viewer = ["py.typed"]


